+++ src/cdb/db.rs
@@ -1,11 +1,11 @@
 use {
     super::{
         binary::{BinaryLinkIF, BinaryLinkList},
+        clause::{Clause, ClauseIF},
         ema::ProgressLBD,
         property,
         watch_cache::*,
-        BinaryLinkDB, CertificationStore, Clause, ClauseDB, ClauseDBIF, ClauseId, ReductionType,
-        RefClause,
+        BinaryLinkDB, CertificationStore, ClauseDBIF, ClauseId, ReductionType, RefClause,
     },
     crate::{assign::AssignIF, types::*},
     std::{
@@ -17,6 +17,82 @@ use {
 
 #[cfg(not(feature = "no_IO"))]
 use std::{fs::File, io::Write, path::Path};
+
+/// use crate::splr::cdb::ClauseDB;
+/// let cdb = ClauseDB::instantiate(&Config::default(), &CNFDescription::default());
+///```
+#[derive(Clone, Debug)]
+pub struct ClauseDB {
+    /// container of clauses
+    pub(super) clause: Vec<Clause>,
+    /// hashed representation of binary clauses.
+    ///## Note
+    /// This means a biclause \[l0, l1\] is stored at bi_clause\[l0\] instead of bi_clause\[!l0\].
+    ///
+    binary_link: BinaryLinkDB,
+    /// container of watch literals
+    pub(super) watch_cache: Vec<WatchCache>,
+    /// collected free clause ids.
+    freelist: Vec<ClauseId>,
+    /// see unsat_certificate.rs
+    certification_store: CertificationStore,
+    /// a number of clauses to emit out-of-memory exception
+    soft_limit: usize,
+    /// 'small' clause threshold
+    co_lbd_bound: u16,
+    // not in use
+    // lbd_frozen_clause: usize,
+
+    // bi-clause completion
+    bi_clause_completion_queue: Vec<Lit>,
+    pub(super) num_bi_clause_completion: usize,
+
+    //
+    //## clause rewarding
+    //
+    /// an index for counting elapsed time
+    #[cfg(feature = "clause_rewarding")]
+    tick: usize,
+    #[cfg(feature = "clause_rewarding")]
+    activity_decay: f64,
+    #[cfg(feature = "clause_rewarding")]
+    activity_anti_decay: f64,
+
+    //
+    //## LBD
+    //
+    /// a working buffer for LBD calculation
+    lbd_temp: Vec<usize>,
+    pub(super) lbd: ProgressLBD,
+
+    //
+    //## statistics
+    //
+    /// the number of active (not DEAD) clauses.
+    pub(super) num_clause: usize,
+    /// the number of binary clauses.
+    pub(super) num_bi_clause: usize,
+    /// the number of binary learnt clauses.
+    pub(super) num_bi_learnt: usize,
+    /// the number of clauses which LBDs are 2.
+    pub(super) num_lbd2: usize,
+    /// the present number of learnt clauses.
+    pub(super) num_learnt: usize,
+    /// the number of reductions.
+    pub(super) num_reduction: usize,
+    /// the number of reregistration of a bi-clause
+    pub(super) num_reregistration: usize,
+    /// Literal Block Entanglement
+    /// EMA of LBD of clauses used in conflict analysis (dependency graph)
+    pub(super) lb_entanglement: Ema2,
+    /// cutoff value used in the last `reduce`
+    pub(super) reduction_threshold: f64,
+
+    //
+    //## incremental solving
+    //
+    pub eliminated_permanent: Vec<Vec<Lit>>,
+}
 
 impl Default for ClauseDB {
     fn default() -> ClauseDB {
@@ -440,6 +516,7 @@ impl ClauseDBIF for ClauseDB {
             cid,
             c,
         );
+        self.freelist.push(cid);
         // assert_eq!(self.clause.iter().skip(1).filter(|c| !c.is_dead()).count(), self.num_clause);
     }
     fn remove_clause_sandbox(&mut self, cid: ClauseId) {
@@ -461,6 +538,7 @@ impl ClauseDBIF for ClauseDB {
             cid,
             c,
         );
+        self.freelist.push(cid);
         // assert_eq!(self.clause.iter().skip(1).filter(|c| !c.is_dead()).count(), self.num_clause);
     }
     //
@@ -935,7 +1013,7 @@ impl ClauseDBIF for ClauseDB {
         if learnt {
             #[cfg(feature = "just_used")]
             c.turn_on(FlagClause::USED);
-            #[cfg(feature = "clause_rewading")]
+            #[cfg(feature = "clause_rewarding")]
             self.reward_at_analysis(cid);
         }
         if 1 < rank {
@@ -945,14 +1023,6 @@ impl ClauseDBIF for ClauseDB {
     }
     /// reduce the number of 'learnt' or *removable* clauses.
     fn reduce(&mut self, asg: &mut impl AssignIF, setting: ReductionType) {
-        impl Clause {
-            fn reverse_activity_sum(&self, asg: &impl AssignIF) -> f64 {
-                self.iter().map(|l| 1.0 - asg.activity(l.vi())).sum()
-            }
-            fn lbd(&self) -> f64 {
-                self.rank as f64
-            }
-        }
         let ClauseDB {
             ref mut clause,
             ref mut lbd_temp,
@@ -1331,5 +1401,11 @@ impl Clause {
             }
         }
         falsified
+    }
+    fn reverse_activity_sum(&self, asg: &impl AssignIF) -> f64 {
+        self.iter().map(|l| 1.0 - asg.activity(l.vi())).sum()
+    }
+    fn lbd(&self) -> f64 {
+        self.rank as f64
     }
 }